refactor: add subtask meta register
