+++ backend/helpers/pluginhelper/api/connection_auths.go
@@ -24,9 +24,8 @@ import (
 	"reflect"
 	"strings"
 
-	"github.com/apache/incubator-devlake/core/plugin"
-
 	"github.com/apache/incubator-devlake/core/errors"
+	"github.com/apache/incubator-devlake/core/plugin"
 	"github.com/go-playground/validator/v10"
 )
