+++ backend/helpers/pluginhelper/subtaskmeta/sorter/subtask.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package subtaskmeta_sorter
+package sorter
 
 import (
 	"fmt"
